+++ packages/contracts/contracts/factory/DAOFactory.sol
@@ -42,11 +42,15 @@ contract DAOFactory {
         uint64 minDuration;
     }
 
+    /// @notice Emitted when a new DAO is created
+    /// @param name The DAO name
+    /// @param token ERC20 DAO token address or address(0) no token was created
+    /// @param voting The address of the voting component of the new DAO
     event DAOCreated(string name, address indexed token, address indexed voting);
 
-    // @dev Stores the registry and token factory address and creates the base contracts required for the factory
-    // @param _registry The DAO registry to register the DAO with his name
-    // @param _tokenFactory The Token Factory to register tokens
+    /// @dev Stores the registry and token factory address and creates the base contracts required for the factory
+    /// @param _registry The DAO registry to register the DAO with his name
+    /// @param _tokenFactory The Token Factory to register tokens
     constructor(Registry _registry, TokenFactory _tokenFactory) {
         registry = _registry;
         tokenFactory = _tokenFactory;
@@ -54,12 +58,12 @@ contract DAOFactory {
         setupBases();
     }
 
-    // @dev Creates a new DAO with ERC20 based voting. It also will deploy a new token if the corresponding config is passed.
-    // @oaram _daoConfig The name and metadata hash of the DAO it creates
-    // @param _voteConfig The majority voting configs and minimum duration of voting
-    // @param _tokenConfig The token config used to deploy a new token
-    // @param _mintConfig The config for the minter for the newly created token
-    // @param _gsnForwarder The forwarder address for the OpenGSN meta tx solution
+    /// @dev Creates a new DAO with ERC20 based voting. It also will deploy a new token if the corresponding config is passed.
+    /// @param _daoConfig The name and metadata hash of the DAO it creates
+    /// @param _voteConfig The majority voting configs and minimum duration of voting
+    /// @param _tokenConfig The token config used to deploy a new token
+    /// @param _mintConfig The config for the minter for the newly created token
+    /// @param _gsnForwarder The forwarder address for the OpenGSN meta tx solution
     function newERC20VotingDAO(
         DAOConfig calldata _daoConfig,
         VoteConfig calldata _voteConfig,
@@ -99,11 +103,11 @@ contract DAOFactory {
         emit DAOCreated(_daoConfig.name, address(token), address(voting));
     }
 
-    // @dev Creates a new DAO with whitelist based voting.
-    // @oaram _daoConfig The name and metadata hash of the DAO it creates
-    // @param _voteConfig The majority voting configs and minimum duration of voting
-    // @param _whitelisVoters A array of addresses that are allowed to vote
-    // @param _gsnForwarder The forwarder address for the OpenGSN meta tx solution
+    /// @dev Creates a new DAO with whitelist based voting.
+    /// @param _daoConfig The name and metadata hash of the DAO it creates
+    /// @param _voteConfig The majority voting configs and minimum duration of voting
+    /// @param _whitelistVoters An array of addresses that are allowed to vote
+    /// @param _gsnForwarder The forwarder address for the OpenGSN meta tx solution
     function newWhitelistVotingDAO(
         DAOConfig calldata _daoConfig,
         VoteConfig calldata _voteConfig,
@@ -122,9 +126,9 @@ contract DAOFactory {
         emit DAOCreated(_daoConfig.name, address(0), address(voting));
     }
 
-    // @dev Creates a new DAO.
-    // @oaram _daoConfig The name and metadata hash of the DAO it creates
-    // @param _gsnForwarder The forwarder address for the OpenGSN meta tx solution
+    /// @dev Creates a new DAO.
+    /// @param _daoConfig The name and metadata hash of the DAO it creates
+    /// @param _gsnForwarder The forwarder address for the OpenGSN meta tx solution
     function createDAO(DAOConfig calldata _daoConfig, address _gsnForwarder)
         internal
         returns (DAO dao)
@@ -135,9 +139,9 @@ contract DAOFactory {
         dao.initialize(_daoConfig.metadata, address(this), _gsnForwarder);
     }
 
-    // @dev Does set the required permissions for the new DAO.
-    // @param _dao The DAO instance just created.
-    // @param _voting The voting contract address (whitelist OR ERC20 voting)
+    /// @dev Does set the required permissions for the new DAO.
+    /// @param _dao The DAO instance just created.
+    /// @param _voting The voting contract address (whitelist OR ERC20 voting)
     function setDAOPermissions(DAO _dao, address _voting) internal {
         // set roles on the dao itself.
         ACLData.BulkItem[] memory items = new ACLData.BulkItem[](8);
@@ -165,7 +169,10 @@ contract DAOFactory {
         _dao.bulk(address(_dao), items);
     }
 
-    /// @dev internal helper method to create ERC20Voting
+    /// @dev Internal helper method to create ERC20Voting
+    /// @param _dao The DAO address
+    /// @param _token The ERC20 Upgradeable token address
+    /// @param _voteConfig The ERC20 voting settings for the DAO
     function createERC20Voting(
         DAO _dao,
         ERC20VotesUpgradeable _token,
@@ -207,7 +214,10 @@ contract DAOFactory {
         _dao.bulk(address(erc20Voting), items);
     }
 
-    /// @dev internal helper method to create Whitelist Voting
+    /// @dev Internal helper method to create Whitelist Voting
+    /// @param _dao The DAO address creating the Whitelist Voting
+    /// @param _whitelistVoters The array of the allowed voting addresses
+    /// @param _voteConfig The voting settings
     function createWhitelistVoting(
         DAO _dao,
         address[] calldata _whitelistVoters,
@@ -254,7 +264,7 @@ contract DAOFactory {
         _dao.bulk(address(whitelistVoting), items);
     }
 
-    // @dev Internal helper method to set up the required base contracts on DAOFactory deployment.
+    /// @dev Internal helper method to set up the required base contracts on DAOFactory deployment.
     function setupBases() private {
         erc20VotingBase = address(new ERC20Voting());
         whitelistVotingBase = address(new WhitelistVoting());